New 'normaliser_sources' functionality.
Normalisers are now specified separately from the data_sources, this was done for a few reasons,
- Normalisers can be loaded with no data_source (e.g. for loading data-set level info)
- Normaliser specification does not need to be repeated in train/valid/test data_source specification
- Avoids the need to create a "switch" statement in data.py normalisation logic

Additionally, a ValueError check in FilesDataset was fixed. Previously it was assumed that all normalisers had a corresponding data_source. Now this statement checks if the data_source exists before checking consistency of 'use_deltas' setup.